feat: Implement user reviews and enhance UI/UX
This commit introduces a user review system, allowing users to view ratings and reviews for fishers and buyers. It also includes significant UI/UX enhancements, such as changing price fields to integers, updating app branding, and improving navigation feedback.

Key changes:
-   **User Reviews & Ratings:**
    -   **New Screens (`fisher_review_screen.dart`, `buyer_review_screen.dart`):** Created dedicated screens to display a user's average rating, a rating distribution summary, and a list of individual reviews.
    -   **Navigation:** Users can now tap on a seller/buyer card in the `OfferDetails` screen to navigate to their review page.
    -   **Animated Ratings:** Replaced static star icons with the `animated_rating_stars` package for a more dynamic and interactive rating display in `RatingCard`, `ReviewCard`, and the rating modal.
    -   **Routing:** Added new routes (`/reviews/:userId`) for both buyer and fisher roles to handle navigation to review screens.

-   **Data & State Management:**
    -   **Integer Pricing:** Refactored data models (`Catch`, `Offer`), repositories, BLoCs, and UI components to use `int` for all price-related fields (`price`, `pricePerKg`, `total`) instead of `double`. This simplifies input and display.
    -   **Number Input Fields:** Updated `NumberInputField` and dialogs to enforce integer-only input for prices, preventing decimal entries.
    -   **Notification Badges:**
        -   Added notification counts to the "Offers" and "Messages" tabs on the `NotificationsScreen` to indicate pending updates.
        -   Implemented a badge on the `CatchDetails` screen's "Offers" tab to show the number of offers needing attention.

-   **UI/UX Enhancements:**
    -   **App Branding:**
        -   Updated the Android application label to "Siren Marketplace".
        -   Replaced all old `ic_launcher` (Android) and `AppIcon` (iOS) assets with new, branded application icons.
    -   **Page Titles:** Created a reusable `PageTitle` widget and applied it across multiple screens (`OfferDetails`, `CatchDetails`, `NotificationsScreen`, `AccountInfo`) for consistent styling.
    -   **Clickable User Cards:** The seller/buyer profile section in `OfferDetails` is now a clickable card that navigates to their reviews.
    -   **Catch Details Display:** Improved the `CatchDetails` and `ForSaleCard` to correctly display "Size" for prawns and "Average Size" in cm for other species.

-   **Dependencies:**
    -   Added the `animated_rating_stars: ^1.0.1` package for improved rating visualization.

+++ lib/core/models/offer.dart
@@ -9,14 +9,14 @@ class Offer extends Equatable {
   final String catchId;
   final String fisherId;
   final String buyerId;
-  final double pricePerKg;
-  final double price;
+  final int pricePerKg;
+  final int price;
   final double weight;
   final OfferStatus status;
   final String dateCreated;
-  final double? previousPrice;
+  final int? previousPrice;
   final double? previousWeight;
-  final double? previousPricePerKg;
+  final int? previousPricePerKg;
 
   // 🆕 ADDED FIELD
   final bool hasUpdateForFisher;
@@ -95,12 +95,12 @@ class Offer extends Equatable {
   // Used for transactional updates (e.g., accepting or countering)
   Offer copyWith({
     OfferStatus? status,
-    double? previousPrice,
+    int? previousPrice,
     double? previousWeight,
-    double? previousPricePerKg,
-    double? price,
+    int? previousPricePerKg,
+    int? price,
     double? weight,
-    double? pricePerKg,
+    int? pricePerKg,
     String? dateCreated,
     bool? hasUpdateForFisher,
     bool? hasUpdateForBuyer,
@@ -183,14 +183,14 @@ class Offer extends Equatable {
     catchId: m['catch_id'] as String,
     fisherId: m['fisher_id'] as String,
     buyerId: m['buyer_id'] as String,
-    pricePerKg: (m['price_per_kg'] as num).toDouble(),
-    price: (m['price'] as num).toDouble(),
+    pricePerKg: (m['price_per_kg'] as num).toInt(),
+    price: (m['price'] as num).toInt(),
     weight: (m['weight'] as num).toDouble(),
     status: offerStatusFromString(m['status'] as String),
     dateCreated: m['date_created'] as String,
-    previousPrice: (m['previous_price'] as num?)?.toDouble(),
+    previousPrice: (m['previous_price'] as num?)?.toInt(),
     previousWeight: (m['previous_weight'] as num?)?.toDouble(),
-    previousPricePerKg: (m['previous_price_per_kg'] as num?)?.toDouble(),
+    previousPricePerKg: (m['previous_price_per_kg'] as num?)?.toInt(),
     // 🆕 EXTRACTED FIELD
     hasUpdateForBuyer: (m['has_update_buyer'] as int) == 1,
     hasUpdateForFisher: (m['has_update_fisher'] as int) == 1,
